requirements.txt
Flask==0.10.1

mysql-connector==2.2.9


